reorganize directories; rename files to camelCase format
