+++ src/public/js/config/appConfig.js
